hotfix: fix scripts
